removed the underscore from the inference submodules
