+++ probeye/inference/__init__.py
@@ -1,5 +1,5 @@
 # subpackage imports
-from probeye.inference import scipy_
-from probeye.inference import emcee_
-from probeye.inference import dynesty_
-from probeye.inference import torch_
+from probeye.inference import scipy
+from probeye.inference import emcee
+from probeye.inference import dynesty
+from probeye.inference import torch
