+++ probeye/inference/dynesty/__init__.py
@@ -1,0 +1,2 @@
+# module imports
+from probeye.inference.dynesty import solver
