+++ probeye/inference/emcee/__init__.py
@@ -1,0 +1,2 @@
+# module imports
+from probeye.inference.emcee import solver
