--- probeye/inference/emcee_/__init__.py
@@ -1,2 +1,0 @@
-# module imports
-from probeye.inference.emcee_ import solver
