+++ probeye/inference/scipy/__init__.py
@@ -1,0 +1,5 @@
+# module imports
+from probeye.inference.scipy import correlation_models
+from probeye.inference.scipy import likelihood_models
+from probeye.inference.scipy import priors
+from probeye.inference.scipy import solver
