+++ probeye/inference/scipy/priors.py
