--- probeye/inference/scipy_/__init__.py
@@ -1,5 +1,0 @@
-# module imports
-from probeye.inference.scipy_ import correlation_models
-from probeye.inference.scipy_ import likelihood_models
-from probeye.inference.scipy_ import priors
-from probeye.inference.scipy_ import solver
