+++ probeye/inference/torch/__init__.py
@@ -1,0 +1,4 @@
+# module imports
+from probeye.inference.torch import likelihood_models
+from probeye.inference.torch import priors
+from probeye.inference.torch import solver
