--- probeye/inference/torch_/__init__.py
@@ -1,4 +1,0 @@
-# module imports
-from probeye.inference.torch_ import likelihood_models
-from probeye.inference.torch_ import priors
-from probeye.inference.torch_ import solver
