+++ tests/integration_tests/tests_with_correlation/test_time_correlation.py
@@ -69,8 +69,8 @@ class TestProblem(unittest.TestCase):
             If True, the problem is solved with the emcee solver. Otherwise,
             the emcee solver will not be used.
         run_torch : bool, optional
-            If True, the problem is solved with the pyro/torch_ solver.
-            Otherwise, the pyro/torch_ solver will not be used.
+            If True, the problem is solved with the pyro/torch solver.
+            Otherwise, the pyro/torch solver will not be used.
         run_dynesty
             If True, the problem is solved with the dynesty solver. Otherwise, the
             dynesty solver will not be used.
