+++ tests/integration_tests/tests_without_correlation/test_maximum_likelihood.py
@@ -21,7 +21,7 @@ from probeye.definition.sensor import Sensor
 from probeye.definition.likelihood_model import GaussianLikelihoodModel
 
 # local imports (inference related)
-from probeye.inference.scipy_.solver import ScipySolver
+from probeye.inference.scipy.solver import ScipySolver
 
 
 class TestProblem(unittest.TestCase):
