+++ tests/unit_tests/inference/dynesty/test_solver.py
@@ -10,7 +10,7 @@ from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.likelihood_model import GaussianLikelihoodModel
-from probeye.inference.dynesty_.solver import DynestySolver
+from probeye.inference.dynesty.solver import DynestySolver
 
 
 class TestProblem(unittest.TestCase):
