+++ tests/unit_tests/inference/emcee/test_solver.py
@@ -10,7 +10,7 @@ from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.likelihood_model import GaussianLikelihoodModel
-from probeye.inference.emcee_.solver import EmceeSolver
+from probeye.inference.emcee.solver import EmceeSolver
 
 
 class TestProblem(unittest.TestCase):
