+++ tests/unit_tests/inference/torch/test_solver.py
@@ -10,7 +10,7 @@ from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.likelihood_model import GaussianLikelihoodModel
-from probeye.inference.torch_.solver import PyroSolver
+from probeye.inference.torch.solver import PyroSolver
 
 
 class TestProblem(unittest.TestCase):
